Add paged main activity with menu fragments
Added a host game and join game fragment with simple placeholder text. These are contained in a view pager which allows swapping between these.
Fragments are created programmatically not declaratively, which may be an issue later.

Added a TODO around hardcoded text strings, will need to change this to a strings.xml reference later although that may be require a workaround (read: hack)